% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hilsa_ancova}
\alias{hilsa_ancova}
\title{Length-Weight ANCOVA Dataset for Hilsa Shad}
\format{
A data frame with 90 rows and 3 variables:
\describe{
  \item{Habitat}{A factor representing the environment (Marine, Estuary, River).}
  \item{Total_Length_cm}{A numeric vector representing total fish length in centimeters (covariate).}
  \item{Weight_g}{A numeric vector representing fish body weight in grams (response).}
}
}
\usage{
hilsa_ancova
}
\description{
A simulated dataset containing body weight, total length, and habitat types
of Hilsa shad (Tenualosa ilisha) to demonstrate Analysis of Covariance (ANCOVA).
}
\keyword{datasets}
