% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hilsa_catch}
\alias{hilsa_catch}
\title{Catch Count Dataset for Hilsa Shad GLMs}
\format{
A data frame with 120 rows and 4 variables:
\describe{
  \item{Habitat}{A factor representing the environment (Marine, Estuary, River).}
  \item{Season}{A factor representing the season (Monsoon, Dry).}
  \item{Fishing_Hours}{A numeric vector representing fishing effort in hours.}
  \item{Catch_Count}{An integer vector representing the number of fish caught (count response).}
}
}
\usage{
hilsa_catch
}
\description{
A simulated dataset containing catch counts, fishing effort, and environmental
variables for Hilsa shad (Tenualosa ilisha) to demonstrate Generalized Linear Models (GLMs).
}
\keyword{datasets}
