% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hilsa_env}
\alias{hilsa_env}
\title{Environmental and Catch Parameters for Correlation Analysis}
\format{
A data frame with 100 rows and 6 variables:
\describe{
  \item{SST_C}{Sea Surface Temperature in degrees Celsius.}
  \item{Salinity_ppt}{Water salinity in parts per thousand.}
  \item{Depth_m}{Water depth in meters.}
  \item{DO_mgL}{Dissolved oxygen in milligrams per liter.}
  \item{Chlorophyll_a}{Chlorophyll-a concentration.}
  \item{Catch_kg}{Total fish catch in kilograms.}
}
}
\usage{
hilsa_env
}
\description{
A simulated dataset containing oceanographic measurements and catch weights
to demonstrate customizable correlation matrix heatmaps.
}
\keyword{datasets}
