% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hilsa_morphology}
\alias{hilsa_morphology}
\title{Morphometric Measurements of Hilsa Shad}
\format{
A data frame with 90 rows and 4 variables:
\describe{
  \item{Habitat}{A factor representing the environment (Marine, Estuary, River).}
  \item{Body_Depth_cm}{A numeric vector representing maximum body depth in centimeters.}
  \item{Head_Length_cm}{A numeric vector representing head length in centimeters.}
  \item{Fin_Length_cm}{A numeric vector representing pectoral fin length in centimeters.}
}
}
\usage{
hilsa_morphology
}
\description{
A simulated dataset containing multiple continuous morphological traits
of Hilsa shad (Tenualosa ilisha) across different aquatic habitats to
demonstrate Multivariate Analysis of Variance (MANOVA).
}
\keyword{datasets}
