% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hilsa_two_way}
\alias{hilsa_two_way}
\title{Multi-Factor Body Weights of Hilsa Shad}
\format{
A data frame with 90 rows and 3 variables:
\describe{
  \item{Habitat}{A factor representing the environment (Marine, Estuary, River).}
  \item{Season}{A factor representing the season (Monsoon, Dry).}
  \item{Weight_g}{A numeric vector representing fish body weight in grams.}
}
}
\usage{
hilsa_two_way
}
\description{
A simulated dataset containing the body weights of Hilsa shad
(Tenualosa ilisha) across different habitats and fishing seasons
to demonstrate two-way ANOVA interaction effects.
}
\keyword{datasets}
