% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hilsa_weight}
\alias{hilsa_weight}
\title{Body Weights of Hilsa Shad Across Habitats}
\format{
A data frame with 90 rows and 2 variables:
\describe{
\item{Habitat}{A factor representing the sampling environment (Marine, Estuary, River).}
\item{Weight_g}{A numeric vector representing the body weight of the fish in grams.}
}
}
\usage{
hilsa_weight
}
\description{
A simulated dataset containing the body weights of Hilsa shad
(Tenualosa ilisha) sampled from three distinct ecological environments:
Marine, Estuary, and River. This dataset is designed to demonstrate
one-way ANOVA and variance visualization.
}
\keyword{datasets}
